Long quantity

+++ src/main/java/ait/cohort46/student/service/StudentService.java
@@ -21,7 +21,7 @@ public interface StudentService {
 
     List<StudentDto> findStudentByName(String name);
 
-    Integer getStudentQuantityByNames(Set<String> names);
+    Long getStudentQuantityByNames(Set<String> names);
 
     List<StudentDto> findStudentsByExamMinScore(String examName, Integer minScore);
 }
